refactor: split services and options
