+++ Bot.Core/Services/ConversationStateOptions.cs
@@ -1,0 +1,7 @@
+namespace Bot.Core.Services;
+
+public class ConversationStateOptions
+{
+    public string RedisKeyPrefix { get; set; } = "session:";
+    public TimeSpan SessionTtl { get; set; } = TimeSpan.FromHours(24);
+}
