+++ Bot.Core/Services/ISpeechRecognizer.cs
@@ -1,0 +1,6 @@
+namespace Bot.Core.Services;
+
+public interface ISpeechRecognizer : IAsyncDisposable
+{
+    Task<RecognitionResult> RecognizeOnceAsync();
+}
