+++ Bot.Core/Services/ISpeechService.cs
@@ -1,0 +1,6 @@
+namespace Bot.Core.Services;
+
+public interface ISpeechService
+{
+    Task<(string Text, string DetectedLanguage)> TranscribeAsync(Stream audioStream);
+}
