+++ Bot.Core/Services/ITranscriptionService.cs
@@ -1,0 +1,6 @@
+namespace Bot.Core.Services;
+
+public interface ITranscriptionService
+{
+    Task<(string Text, string DetectedLanguage)> TranscribeAsync(Stream audioStream, string[] languageCodes);
+}
