+++ Bot.Core/Services/RecognitionResult.cs
@@ -1,0 +1,5 @@
+using Microsoft.CognitiveServices.Speech;
+
+namespace Bot.Core.Services;
+
+public record RecognitionResult(ResultReason Reason, string Text, string DetectedLanguage);
