+++ Bot.Core/Services/SpeechService.cs
@@ -4,29 +4,6 @@ using Microsoft.Extensions.Options;
 
 namespace Bot.Core.Services;
 
-/// <summary>
-///     Configuration for speech transcription: which languages to auto-detect.
-/// </summary>
-public class SpeechOptions
-{
-    /// <summary>
-    ///     A list of BCP-47 codes the TranscriptionService will try to auto-detect.
-    ///     E.g. "en-US","ig-NG","yo-NG","ha-NG","pcm-NG" (Nigerian Pidgin).
-    /// </summary>
-    public string[] SupportedLanguages { get; set; } = new[]
-    {
-        "en-US",
-        "ig-NG",
-        "yo-NG",
-        "ha-NG",
-        "pcm-NG"
-    };
-}
-
-public interface ISpeechService
-{
-    Task<(string Text, string DetectedLanguage)> TranscribeAsync(Stream audioStream);
-}
 
 public class SpeechService(
     ITranscriptionService transcriber,
