+++ Bot.Core/Services/WhatsAppService.cs
@@ -5,25 +5,7 @@ using Microsoft.Extensions.Options;
 
 namespace Bot.Core.Services;
 
-public interface IWhatsAppService
-{
-    Task<string> SendTextMessageAsync(string toPhoneNumber, string message);
-    Task<string> SendMediaAsync(string toPhoneNumber, string mediaUrl, string caption = "");
-    Task<string> SendQuickReplyAsync(string toPhoneNumber, string header, string body, string[] buttonLabels);
-    Task<string> SendTemplateAsync(string toPhoneNumber, object template);
-    Task<string> SendVoiceAsync(string toPhoneNumber, Stream audio);
-    Task DeleteMessageAsync(string messageId);
-}
-
 // Bot.Infrastructure/Services/WhatsAppService.cs
-
-public class WhatsAppOptions
-{
-    public string BaseUrl { get; set; } = "https://graph.facebook.com/v18.0";
-    public string PhoneNumberId { get; set; } = null!;
-    public string AccessToken { get; set; } = null!;
-    public TimeSpan EphemeralTtl { get; set; } = TimeSpan.FromMinutes(5);
-}
 
 public class WhatsAppService : IWhatsAppService
 {
