+++ Bot.Tests/Services/ConversationStateServiceTests.cs
@@ -1,4 +1,5 @@
 using Bot.Core.Services;
+using Bot.Core.Models;
 using FluentAssertions;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
